refactor: split large files to comply with max-lines rule
Extract data plane modules from anomaly-investigation.js and chart.js
to bring both files under the 1000-line ESLint limit:

- investigation-data.js: query logic, caching, ID generation
- chart-state.js: state management, navigation, time calculations
